Artikelen toevoegen werkt nu en de namen zijn aangepast voor consistentie
