+++ Aplikassie2/src/main/java/com/rsvier/workshop/controllers/KlantFormulierController.java
@@ -6,7 +6,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.rsvier.workshop.domein.Persoon;
-import com.rsvier.workshop.domein.Persoon.AccountSoort;
 
 @Controller
 @RequestMapping("/klantformulier")
