+++ Aplikassie2/src/main/java/com/rsvier/workshop/controllers/RegisterArtikelController.java
@@ -1,0 +1,27 @@
+package com.rsvier.workshop.controllers;
+
+import com.rsvier.workshop.dao.ArtikelRepository;
+import com.rsvier.workshop.domein.Artikel;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/artikeltoegevoegd")
+public class RegisterArtikelController {
+    CrudRepository repository;
+	
+    @Autowired
+    public RegisterArtikelController(ArtikelRepository artikelRepository) {
+        repository = artikelRepository;
+    }
+
+    @PostMapping
+    public Artikel registerKlant(Artikel artikel) {
+        artikel.setArtikelStatus(Artikel.ArtikelStatus.ACTIEF);
+        repository.save(artikel);
+        return artikel;
+    }
+}
