+++ Aplikassie2/src/main/java/com/rsvier/workshop/domein/Artikel.java
@@ -3,10 +3,13 @@ package com.rsvier.workshop.domein;
 import java.math.BigDecimal;
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import org.hibernate.validator.constraints.NotBlank;
 
 @Entity
 @Table(name="artikel")
@@ -25,12 +28,13 @@ public class Artikel {
     @Column(nullable = false)
     protected int voorraad;
     
-    @Column(nullable = false)
-    protected boolean actief;
+    @Enumerated(EnumType.STRING)
+    @Column(name="artikel_status", nullable=false)
+    protected ArtikelStatus artikelStatus;
     
-    public Artikel(boolean actief) {
-        this.actief = actief;
-    }
+    public Artikel() {}
+    
+    public enum ArtikelStatus {ACTIEF, INACTIEF}
     
     public Long getId() {
         return id;
@@ -64,12 +68,12 @@ public class Artikel {
         this.voorraad = voorraad;
     }
 
-    public boolean isActief() {
-        return actief;
+    public ArtikelStatus getArtikelStatus() {
+        return artikelStatus;
     }
 
-    public void setActief(boolean actief) {
-        this.actief = actief;
+    public void setArtikelStatus(ArtikelStatus artikelStatus) {
+        this.artikelStatus = artikelStatus;
     }
 
     @Override
