+++ Aplikassie2/src/main/resources/templates/artikelformulier.html
@@ -1,0 +1,28 @@
+<!DOCTYPE html>
+<html xmlns="http://www.w3.org/1999/xhtml" 
+      xmlns:th="http://www.thymeleaf.org">
+    <head>
+        <title>Artikel toevoegen</title>
+    </head>
+    <body>
+        <h1>Voeg een artikel toe</h1>
+        <form method="POST" action="/artikeltoegevoegd" th:action="@{/artikeltoegevoegd}" th:object="${artikel}">
+            <p> <label for="naam">Naam:</label>
+            <input type="text" th:field="*{naam}" />
+            <span class="validationError"
+            th:if="${#fields.hasErrors('naam')}"
+            th:errors="*{naam}">Naam Error</span></p>
+            <p> <label for="prijs">Prijs:</label>
+            <input type="text" th:field="*{prijs}" />
+            <span class="validationError"
+            th:if="${#fields.hasErrors('prijs')}"
+            th:errors="*{prijs}">Prijs Error</span></p>
+            <p><label for="voorraad">Voorraad:</label>
+            <input type="text" th:field="*{voorraad}" />
+            <span class="validationError"
+            th:if="${#fields.hasErrors('voorraad')}"
+            th:errors="*{voorraad}">Voorraad Error</span></p>
+            <p><input type="submit" value="Voeg toe" /> <input type="reset" value="Reset" /></p>
+	</form>
+    </body>
+</html>
