--- Aplikassie2/src/main/resources/templates/artikeltoevoegen.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" 
-      xmlns:th="http://www.thymeleaf.org">
-    <head>
-        <title>Artikel toevoegen</title>
-    </head>
-    <body>
-        <h1>Voeg een artikel toe</h1>
-        <form method="POST" action="/register" th:action="@{/registreren}" th:object="${artikel}">
-            <p> <label for="naam">Naam:</label>
-            <input type="text" th:field="*{naam}" /></p>
-            <p> <label for="prijs">Prijs:</label>
-            <input type="text" th:field="*{prijs}" /></p>
-            <p><label for="voorraad">Voorraad:</label>
-            <input type="text" th:field="*{voorraad}" /></p>
-            <p><input type="submit" value="Voeg toe" /> <input type="reset" value="Reset" /></p>
-	</form>
-    </body>
-</html>
